+++ src/components/Common/Header/HeaderLogo.style.jsx
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled'
+
+export const LogoContainer = styled.div`
+  z-index: 1000;
+  position: absolute;
+`
